+++ utilities/inc/gutenberg/inc/accordion/accordion.php
@@ -6,7 +6,7 @@ define( 'ACCORDION_DIRECTORY_URI', GUTENBERG_DIRECTORY_URI . '/inc/accordion' );
 add_action( 'wp_enqueue_scripts', function() {
 	wp_enqueue_style('accordion-style', ACCORDION_DIRECTORY_URI . '/css/style.css' );
 	wp_enqueue_script('accordion-app', ACCORDION_DIRECTORY_URI . '/js/app.js', [], false, true );
-} );
+}, 9 );
 
 /**
  * Register block and enqueue styles and scripts.